update queries and types

+++ ClickHouse.Core/Interfaces/Repositories/ISensorRepository.cs
@@ -7,8 +7,8 @@ public interface ISensorRepository
     Task<IEnumerable<int>> GetLocationsAsync();
     Task<LocationSample> GetMedianForLocationAndTimeRangeAsync(int locationId);
     Task<DateTime> GetLatestSensorTimeAsync();
-    Task<int> GetTotalSensorsCountAsync();
-    Task<int> GetTotalSamplesCountAsync();
+    Task<ulong> GetTotalSensorsCountAsync();
+    Task<ulong> GetTotalSamplesCountAsync();
     Task<IEnumerable<DateCount>> GetSampleCountsPerDateAsync();
     Task<IEnumerable<SensorCount>> GetSampleCountsPerSensorAsync();
 }
