+++ ClickHouse.Core/Models/DemoStats.cs
@@ -3,6 +3,6 @@ namespace ClickHouse.Core.Models;
 public class DemoStats
 {
     public DateTime LatestSensorTime { get; set; }
-    public int TotalSensors { get; set; }
-    public int TotalSamples { get; set; }
+    public ulong TotalSensors { get; set; }
+    public ulong TotalSamples { get; set; }
 }
